Split out unit tests from the Client Library project
