+++ ClientLibrary.Test/Properties/AssemblyInfo.cs
@@ -1,0 +1,21 @@
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+
+[assembly: AssemblyTitle("Amazon Kinesis Client Library .NET Test")]
+[assembly: AssemblyDescription("Amazon Kinesis Client Library for .NET Test")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Amazon.com, Inc")]
+[assembly: AssemblyProduct("Amazon Kinesis Client Library .NET")]
+[assembly: AssemblyCopyright("Copyright 2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: ComVisible(false)]
+
+[assembly: Guid("40800b0a-faf7-4de2-92c6-2057a5901e71")]
+
+// [assembly: AssemblyVersion("1.0.*")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
