+++ ClientLibrary/Properties/AssemblyInfo.cs
@@ -28,3 +28,4 @@ using System.Runtime.CompilerServices;
 
 // Required for Substitutes in tests
 [assembly:InternalsVisibleTo("DynamicProxyGenAssembly2")]
+[assembly:InternalsVisibleTo("ClientLibrary.Test")]
