+++ ClientLibrary/packages.config
@@ -1,6 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NSubstitute" version="1.8.1.0" targetFramework="net45" />
-  <package id="NUnit" version="2.6.4" targetFramework="net45" />
   <package id="Stateless" version="2.5.11.0" targetFramework="net45" />
 </packages>
